added a system to pick which object in the scene you want to be the focal point.

+++ CS4800Assignment1/Assets/Scripts/Mug.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Mug : MonoBehaviour
 {
     [SerializeField]
-    int offSet;
+    float offSet;
 
     const int tileWidth = 4;
     const int tileHeight = 4;
